# SmartHelpDeskSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Helpdesk CRM — Swagger đầy đủ",
    "version": "1.0.0",
    "description": "Đối chiếu source ngày 2026-09-12. 25 handler nghiệp vụ / 29 method-path (4 alias auth). JSON UTF-8; UUID dạng chuỗi; Instant ISO-8601 UTC. ResponseGeneral bỏ field null ở cấp envelope. HTTP 201 vẫn có body.status=200. Lỗi Spring Security filter không bảo đảm envelope ApiError. AI chạy RuleBasedLlmAdapter nội bộ, không gọi API LLM bên ngoài."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local; đổi host/port theo môi trường"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng nhập",
        "operationId": "login",
        "description": "Quyền: Public. Trả accessToken và refreshToken. Access mặc định 86400000 ms, refresh gấp 7 lần; phụ thuộc cấu hình.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Login successful",
                  "data": {
                    "accessToken": "<accessToken>",
                    "refreshToken": "<refreshToken>",
                    "user": {
                      "id": "0198f101-1234-7000-8000-000000000001",
                      "email": "client@example.com",
                      "name": "Nguyễn Văn An",
                      "role": "ADMIN",
                      "phone": "Ví dụ",
                      "companyId": "COMPANY001",
                      "companyName": "Công ty mẫu",
                      "status": "active"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "client@example.com",
                "password": "Example123!"
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng ký tài khoản",
        "operationId": "register",
        "description": "Quyền: Public. Email đã tồn tại trả BAD_400. companyId không tồn tại trả RES_404.",
        "x-access": "Public",
        "responses": {
          "201": {
            "description": "Tạo thành công; body.status=200.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registerResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Member registered successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "email": "client@example.com",
                    "name": "Nguyễn Văn An",
                    "role": "ADMIN",
                    "phone": "Ví dụ",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "status": "active",
                    "isDeleted": false,
                    "createdAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "client@example.com",
                "password": "Example123!",
                "name": "Nguyễn Văn An",
                "role": "CLIENT",
                "phone": "0900000000",
                "companyId": "COMPANY001"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Làm mới token",
        "operationId": "refresh",
        "description": "Quyền: Public. Refresh token chỉ dùng một lần; xoay vòng cả hai token. Store trong bộ nhớ nên mất khi restart.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/refreshResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Token refreshed successfully",
                  "data": {
                    "accessToken": "<accessToken>",
                    "refreshToken": "<refreshToken>",
                    "accessTokenTtl": 86400000,
                    "refreshTokenTtl": 604800000
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshRequest"
              },
              "example": {
                "refreshToken": "<refreshToken>"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Thiếu/rỗng dùng UNKNOWN_DEVICE. Trình duyệt tự quản lý header này."
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng xuất",
        "operationId": "logout",
        "description": "Quyền: Public. Không body. Có bearer hợp lệ: thu hồi toàn bộ refresh token của user; không thu hồi access token. Không gửi bearer vẫn trả thành công. Response không có data.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logoutResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Logout successful"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng nhập",
        "operationId": "loginV1",
        "description": "Quyền: Public. Trả accessToken và refreshToken. Access mặc định 86400000 ms, refresh gấp 7 lần; phụ thuộc cấu hình.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginV1Response"
                },
                "example": {
                  "status": 200,
                  "message": "Login successful",
                  "data": {
                    "accessToken": "<accessToken>",
                    "refreshToken": "<refreshToken>",
                    "user": {
                      "id": "0198f101-1234-7000-8000-000000000001",
                      "email": "client@example.com",
                      "name": "Nguyễn Văn An",
                      "role": "ADMIN",
                      "phone": "Ví dụ",
                      "companyId": "COMPANY001",
                      "companyName": "Công ty mẫu",
                      "status": "active"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "client@example.com",
                "password": "Example123!"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng ký tài khoản",
        "operationId": "registerV1",
        "description": "Quyền: Public. Email đã tồn tại trả BAD_400. companyId không tồn tại trả RES_404.",
        "x-access": "Public",
        "responses": {
          "201": {
            "description": "Tạo thành công; body.status=200.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registerV1Response"
                },
                "example": {
                  "status": 200,
                  "message": "Member registered successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "email": "client@example.com",
                    "name": "Nguyễn Văn An",
                    "role": "ADMIN",
                    "phone": "Ví dụ",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "status": "active",
                    "isDeleted": false,
                    "createdAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "client@example.com",
                "password": "Example123!",
                "name": "Nguyễn Văn An",
                "role": "CLIENT",
                "phone": "0900000000",
                "companyId": "COMPANY001"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Làm mới token",
        "operationId": "refreshV1",
        "description": "Quyền: Public. Refresh token chỉ dùng một lần; xoay vòng cả hai token. Store trong bộ nhớ nên mất khi restart.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/refreshV1Response"
                },
                "example": {
                  "status": 200,
                  "message": "Token refreshed successfully",
                  "data": {
                    "accessToken": "<accessToken>",
                    "refreshToken": "<refreshToken>",
                    "accessTokenTtl": 86400000,
                    "refreshTokenTtl": 604800000
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshRequest"
              },
              "example": {
                "refreshToken": "<refreshToken>"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Thiếu/rỗng dùng UNKNOWN_DEVICE. Trình duyệt tự quản lý header này."
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng xuất",
        "operationId": "logoutV1",
        "description": "Quyền: Public. Không body. Có bearer hợp lệ: thu hồi toàn bộ refresh token của user; không thu hồi access token. Không gửi bearer vẫn trả thành công. Response không có data.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logoutV1Response"
                },
                "example": {
                  "status": 200,
                  "message": "Logout successful"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Thông báo của tôi",
        "operationId": "listAlerts",
        "description": "Quyền: Đã đăng nhập. Chỉ targetMemberId của user; mới nhất trước.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/listAlertsResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "unreadOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "true: chỉ chưa đọc; false: tất cả."
          }
        ]
      }
    },
    "/api/alerts/{id}/read": {
      "patch": {
        "tags": [
          "Alerts"
        ],
        "summary": "Đánh dấu thông báo đã đọc",
        "operationId": "readAlert",
        "description": "Quyền: Đúng người nhận thông báo. Không body; gọi lại vẫn thành công. Admin không được đọc thay người khác.",
        "x-access": "Đúng người nhận thông báo",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/readAlertResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Alert marked as read",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "requestId": "0198f101-1234-7000-8000-000000000001",
                    "targetMemberId": "0198f101-1234-7000-8000-000000000001",
                    "alertType": "ASSIGNED",
                    "message": "Ví dụ",
                    "isRead": false,
                    "createdAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Dán accessToken từ login; Authorization: Bearer <accessToken>."
      }
    },
    "schemas": {
...
      "LoginAuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          }
        }
      },
...
      "MemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "password",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "password",
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ],
            "description": "Source hiện chấp nhận cả ADMIN, DEVELOPER, CLIENT trên endpoint đăng ký public."
          },
          "phone": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Tùy chọn; trim khoảng trắng. Nếu khác rỗng phải tồn tại trong companies, nếu không trả 404."
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role"
        ]
      },
...
      "TokenRefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "status",
          "errorCode",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_401",
              "AUTH_403",
              "VAL_400",
              "BAD_400",
              "RES_404",
              "SYS_500"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Lỗi theo field; data null bị bỏ khỏi envelope."
          }
        }
      },
      "loginResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "data": {
            "$ref": "#/components/schemas/LoginAuthResponse"
          }
        }
      },
      "registerResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Member registered successfully"
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        }
      },
      "refreshResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Token refreshed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/RefreshTokenResponse"
          }
        }
      },
      "logoutResponse": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Logout successful"
          }
        }
      },
      "loginV1Response": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "data": {
            "$ref": "#/components/schemas/LoginAuthResponse"
          }
        }
      },
      "registerV1Response": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Member registered successfully"
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        }
      },
      "refreshV1Response": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Token refreshed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/RefreshTokenResponse"
          }
        }
      },
      "logoutV1Response": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Logout successful"
          }
        }
      },
...
      "listAlertsResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            }
          }
        }
      },
      "readAlertResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Alert marked as read"
          },
          "data": {
            "$ref": "#/components/schemas/AlertResponse"
          }
        }
      },
...
    },
    "responses": {
      "Error400": {
        "description": "VAL_400: body/validation sai. BAD_400: nghiệp vụ sai hoặc không có quyền tài nguyên.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 400,
                  "errorCode": "BAD_400",
                  "message": "Bad request"
                }
              },
              "validation": {
                "value": {
                  "status": 400,
                  "errorCode": "VAL_400",
                  "message": "Email cannot be blank",
                  "data": {
                    "email": "Email cannot be blank"
                  }
                }
              }
            }
          }
        }
      },
      "Error401": {
        "description": "Invalid email or password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 401,
                  "errorCode": "AUTH_401",
                  "message": "Invalid email or password"
                }
              }
            }
          }
        }
      },
      "Error404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 404,
                  "errorCode": "RES_404",
                  "message": "Resource not found"
                }
              }
            }
          }
        }
      },
      "Error500": {
        "description": "System internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 500,
                  "errorCode": "SYS_500",
                  "message": "System internal error"
                }
              }
            }
          }
        }
      },
      "SecurityDenied": {
        "description": "403: thiếu/token không hợp lệ hoặc không đủ role ở security filter; body do framework quyết định, có thể rỗng. AccessDeniedException đến controller advice trả ApiError với AUTH_403."
      }
    }
  }
}